Rewrite project to be simpler
The previous version of Jeep was tightly coupled with the JPM module.
This rewrite removes that dependency and instead simply passes commands
to the shell to execute when execution is deferred to JPM.

The previous version also contained more logic within Jeep. This logic
has either been moved into the separate Argy-Bargy library or should now
be contained within the subcommands. As a proof of concept, the dev-deps
subcommand has been rewritten to follow this approach. The other
subcommands have been removed and may be added at a later date.