Slice file name during install to handle path differences on Windows
Jeep declares the `jeep` binscript in its info file using a POSIX-style
relative path. This works fine on POSIX systems because Janet's
`bundle/install` uses the POSIX path separator to slice off the file
name to combine with the binpath. However, this causes an installation
failure on Windows because the path separator there is `\` and not `/`.

The solution is to do the slicing in the `install` function and then
provide the file name as the third argument to `bundle/install`. This
fixes #7 on GitHub Issues.