Make temporary direcory creation more robust
Jeep hard-coded the use of "tmp" as the directory path to use in
functions that needed a temporary directory. This commit uses the
tmp-dir function in util.janet to instead use a temporary directory with
a suffix of eight random capital letters. Five variations are tried
before the function gives up and raises an error.

This fixes #14 in GitHub Issues.