Change how vendored dependencies are specified
Prior to this commit, vendored dependencies were specified in a bundle's
info file in a directory-oriented fashion. That is to say that a
struct/table was associated with the `:vendored` keyword with the keys
of that struct/table representing the directory in which a vendored
dependency would be installed.

This commit changes so that vendored dependencies are specified in a
bundle's info file in a dependency-oriented fashion. As a result,
`:vendored` now maps to an array/tuple that lists the dependencies.
Dependencies specify their location using `:prefix` (as in the past) and
optionally with a new array/tuple value rather than a string.

An array/tuple value should consist of two elements, a source (relative
to the dependency root) and a destination (relative to the prefix). If
no prefix is specified, the destination is relative to the bundle root.
A goal of this new system is to allow for users to vendor files into
different directories than in the directory structure used by the
dependency.

The previous system remains but prints a deprecation warning. There is
no plan to remove it but users are nevertheless encouraged to move to
the new system.

This fixes #15 and #16 in GitHub Issues.